+++ src/main/scala/edu/ie3/simona/ontology/messages/flex/MinMaxFlexOptions.scala
@@ -6,9 +6,19 @@
 
 package edu.ie3.simona.ontology.messages.flex
 
-import edu.ie3.simona.exceptions.CriticalFailureException
-import edu.ie3.util.scala.quantities.DefaultQuantities._
+import edu.ie3.datamodel.models.result.system.FlexOptionsResult
+import edu.ie3.simona.exceptions.{CriticalFailureException, FlexException}
+import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage.{
+  IssueFlexControl,
+  IssueNoControl,
+  IssuePowerControl,
+}
+import edu.ie3.util.quantities.QuantityUtils.*
+import edu.ie3.util.scala.quantities.DefaultQuantities.*
 import squants.Power
+
+import java.time.ZonedDateTime
+import java.util.UUID
 
 /** A [[FlexOptions]] type that provides interval-based flexibility in form of
   * reference, minimum and maximum power of an asset. It is possible that the
@@ -41,7 +51,52 @@ final case class MinMaxFlexOptions(
 
 }
 
-object MinMaxFlexOptions {
+object MinMaxFlexOptions extends FlexOptionsExtra[MinMaxFlexOptions] {
+
+  override val flexType: FlexType = FlexType.MinMax
+
+  override def checkSetPower(
+      flexOptions: MinMaxFlexOptions,
+      setPower: Power,
+  ): Unit = {
+    if (setPower < flexOptions.min)
+      throw new FlexException(
+        s"The set power $setPower must not be lower than the minimum power ${flexOptions.min}!"
+      )
+    else if (setPower > flexOptions.max)
+      throw new FlexException(
+        s"The set power $setPower must not be greater than the maximum power ${flexOptions.max}!"
+      )
+  }
+
+  override def determineFlexPower(
+      flexOptions: MinMaxFlexOptions,
+      flexCtrl: IssueFlexControl,
+  ): Power =
+    flexCtrl match {
+      case IssuePowerControl(_, setPower) =>
+        // sanity check: setPower is in range of latest flex options
+        checkSetPower(flexOptions, setPower)
+
+        setPower
+
+      case IssueNoControl(_) =>
+        // no override, take reference power
+        flexOptions.ref
+    }
+
+  override def createResult(
+      flexOptions: MinMaxFlexOptions,
+      modelUuid: UUID,
+      dateTime: ZonedDateTime,
+  ): FlexOptionsResult =
+    new FlexOptionsResult(
+      dateTime,
+      modelUuid,
+      flexOptions.ref.toMegawatts.asMegaWatt,
+      flexOptions.min.toMegawatts.asMegaWatt,
+      flexOptions.max.toMegawatts.asMegaWatt,
+    )
 
   extension (flexOptions: Iterable[MinMaxFlexOptions]) {
     def flexSum: MinMaxFlexOptions =
@@ -97,4 +152,5 @@ object MinMaxFlexOptions {
       power: Power
   ): MinMaxFlexOptions =
     MinMaxFlexOptions(power, power, power)
+
 }