+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentSpec.scala
@@ -37,7 +37,7 @@ import edu.ie3.simona.model.participant.{
 }
 import edu.ie3.simona.ontology.messages.SchedulerMessage.Completion
 import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage.*
-import edu.ie3.simona.ontology.messages.flex.MinMaxFlexOptions
+import edu.ie3.simona.ontology.messages.flex.{FlexType, MinMaxFlexOptions}
 import edu.ie3.simona.ontology.messages.{
   Activation,
   SchedulerMessage,
@@ -64,7 +64,7 @@ import java.time.temporal.ChronoUnit
   */
 class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
-  private implicit val simulationStartDate: ZonedDateTime =
+  given simulationStartDate: ZonedDateTime =
     TimeUtil.withDefaults.toZonedDateTime("2020-01-01T00:00:00Z")
 
   private val simulationEndDate: ZonedDateTime =
@@ -81,8 +81,11 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
     flexResult = true,
   )
 
-  private implicit val activePowerTolerance: Power = Kilowatts(1e-10)
-  private implicit val reactivePowerTolerance: ReactivePower = Kilovars(1e-10)
+  given FlexType = FlexType.MinMax
+
+  // Testing tolerances
+  given Power = Kilowatts(1e-10)
+  given ReactivePower = Kilovars(1e-10)
 
   "A ParticipantAgent that is not controlled by EM" when {